// src/utils/interfaces.tsx
export interface MoviesInfo {
  adult?: boolean;
  backdrop_path?: string;
  genre_ids?: string[];
  id: string;
  poster_path?: string;
  media_type?: string;
  original_name?: string;
  overview?: string;
  title?: string;
  original_title?: string;
  original_language?: string;
  popularity?: number;
  release_date?: string;
  vote_average?: string;
  video?: boolean;
  vote_count?: string;
  first_air_date?: string;
  origin_country?: string;
}